Add tooltips for Sampling View
This adds a tooltip, which is shown
upon hover over Inclusive and Exclusive
cells in Sampling View.
The tooltip reports the name of the function,
the total number of samples,
the number of samples including the function
(be in on top of the stack or anywhere on the stack
depending on whether the hover is over Inclusive
or Exclusive column), the confidence level of
the confidence interval reported
and the confidence interval itself.
This removes the raw counts from Sampling view.
This removes the percentage sign from the cells
and moves it to the header.

Test: Unit Tests, manually profiled OrbitTest.

Bug: http://b/217363336