Fix crash on Add Iterator click when MS or MA selected
When a user selected a number of dynamically instrumented
functions and a number of manually instrumented scopes,
the "Add Iterator" action is active. A click on was crashing
the client.

The fix consists in a check for the scope_id corresponding
to dynamic instrumentation. The same thing is now done for
"Add Frametrack" aciton. These checks are followed by
ORBIT_CHECK that ensures an InstrumentedFunction instance
is found.

Also, GetScopeId(row) is moved to DataView.

Tests: Manual, Unit
Bug: http://b/232058826