Allow for explicit Typedef arithmetics customization
Currently, Typedef arithmetics relies on the operators defined
for the raw types. In the absence of such, or in case of a need
to overload the behaviour, the only option is to wrap the raw type
into an intermediate type.

This adds customization via adding another template parameter for the
arithmetics-enabling Tags.

Bug: http://b/242038642
Test: Compile, Unit